requirements.txt
amqp==2.5.2
anyjson==0.3.3
asgiref==3.2.3
Babel==2.7.0
billiard==3.6.1.0
celery==4.4.0
Django==3.0.1
django-celery==3.3.1
django-extensions==2.2.5
django-filter==2.2.0
djangorestframework==3.11.0
flower==0.9.3
gunicorn==20.0.4
importlib-metadata==1.3.0
kombu==4.6.7
Markdown==3.1.1
more-itertools==8.0.2
psycopg2==2.8.4
pytz==2019.3
six==1.13.0
sqlparse==0.3.0
tornado==5.1.1
vine==1.3.0
zipp==0.6.0
python-scrapyd-api
Scrapy
scrapyd
